dom/canvas/test/test_offscreencanvas_toblob.html

<!DOCTYPE HTML>
<html>
<head>
<title>WebGL in OffscreenCanvas</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="offscreencanvas.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
</head>
<body>
<canvas id="c" width="64" height="64"></canvas>
<canvas id="c-mt" width="64" height="64"></canvas>
<canvas id="c-ref" width="64" height="64"></canvas>
<script>

SimpleTest.waitForExplicitFinish();

function testBlob(blob, callback) {
  // testing toBlob
  // Fill c-ref with green color.
  var c = document.getElementById("c-ref");
  var ctx = c.getContext("2d");
  ctx.rect(0, 0, 64, 64);
  ctx.fillStyle = "#00FF00";
  ctx.fill();
  var reader = new FileReader();
  reader.onload = function(e) {
    ok(c.toDataURL() == e.target.result, "toBlob should return a 64x64 green square");
    callback();
  };
  reader.readAsDataURL(blob);
}

function runTestOnMainThread() {
  var htmlCanvas = document.getElementById("c-mt");
  ok(htmlCanvas, "Should have HTML canvas element");

  window.onmessage = function(evt) {
    var msg = evt.data || {};
    if (msg.type == "test") {
      ok(msg.result, msg.name);
    }
    if (msg.type == "blob") {
      testBlob(msg.blob, SimpleTest.finish);
    }
  }

  ok(htmlCanvas.transferControlToOffscreen, "HTMLCanvasElement has transferControlToOffscreen function");

  var offscreenCanvas = htmlCanvas.transferControlToOffscreen();
  ok(offscreenCanvas, "Expected transferControlToOffscreen to succeed");

  entryFunction('webgl_toblob', '', offscreenCanvas);
}

function runTest() {

  var htmlCanvas = document.getElementById("c");
  var worker = new Worker("offscreencanvas.js");

  ok(htmlCanvas, "Should have HTML canvas element");
  ok(worker, "Web worker successfully created");

  worker.onmessage = function(evt) {
    var msg = evt.data || {};
    if (msg.type == "test") {
      ok(msg.result, msg.name);
    }
    if (msg.type == "blob") {
      testBlob(msg.blob, function() {
        worker.terminate();
        runTestOnMainThread();
      });
    }
  }

  ok(htmlCanvas.transferControlToOffscreen, "HTMLCanvasElement has transferControlToOffscreen function");

  var offscreenCanvas = htmlCanvas.transferControlToOffscreen();
  ok(offscreenCanvas, "Expected transferControlToOffscreen to succeed");

  worker.postMessage({test: 'webgl_toblob', canvas: offscreenCanvas}, [offscreenCanvas]);
}

SpecialPowers.pushPrefEnv({'set': [
  ['gfx.offscreencanvas.enabled', true],
  ['webgl.force-enabled', true],
]}, runTest);

</script>
</body>
</html>
